Refactor into private module
